backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
cohere==5.0.2
qdrant-client==1.7.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2